Moves federated_user_directory_sync to user_directory handler.
